[3787] Move add scheme extension
