+++ stream-chat-android-ui-utils/src/main/kotlin/io/getstream/chat/android/uiutils/extensions/String.kt
@@ -25,3 +25,14 @@ public fun String.containsLinks(): Boolean {
         """(?:\s|^)((?:https?:)?(?:[a-z|A-Z0-9-]+(?:\.[a-z|A-Z0-9-]+)+)(?::[0-9]+)?(?:(?:[^\s]+)?)?)""".toRegex()
     return this.contains(regex = regex)
 }
+
+/**
+ * Small extension function designed to add a
+ * scheme to URLs that do not have one so that
+ * they can be opened using [android.content.Intent.ACTION_VIEW]
+ */
+public fun String.addSchemeToUrlIfNeeded(): String = when {
+    this.startsWith("http://") -> this
+    this.startsWith("https://") -> this
+    else -> "http://$this"
+}
